Time: 9 ms (98.61%), Space: 41.6 MB (94.11%) - LeetHub

+++ 9-palindrome-number/9-palindrome-number.java
@@ -9,8 +9,7 @@ class Solution {
         while(a>0){
             rem=a%10;
             temp=(temp*10)+rem;
-            a=(int)a/10;
-            
+            a=a/10;  
         }
         if(temp==x){
             return true;
